Add `getCartByCartId` functionality; update port to 8083 in application-docker.yaml
Took 35 minutes

+++ cart-service/src/main/java/org/booknest/cartservice/entity/Cart.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class Cart {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer cartId;
+    private Long cartId;
 
     private Long userId;
